# src/main/resources/static/openapi-pokemon-region.yaml
openapi: 3.1.0
info:
  title: Pokemon Region API
  description: Pokemon Region API allows you to manipulate data from the Pokemon Region database
  version: 1.0.0
  contact:
    name: Murtaza Amjad
    email: murtazaamjad0519@gmail.com
    url: https://murtaza-amjad-personal-website.vercel.app/

paths:
  /region/add:
    post:
      summary: API to add new Pokemon Region
      description: 'The API requires the client to add a new Pokemon Region
      using name and population'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PokemonRegionDTO"
      responses:
        200:
          description: 'Pokemon Region was successfully added'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PokemonRegionDTO"
              examples:
                example-1:
                  $ref: "#/components/schemas/PokemonRegionDTO"
        400:
          description: 'Pokemon Region validation empty exception'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PokemonRegionDTO"
              examples:
                example-1:
                  $ref: "#/components/examples/EmptyRegionExample"
                example-2:
                  $ref: "#/components/examples/MinimumPopulationExample"
  /region/delete/:
    delete:
      summary: API to delete a Pokemon Region
      description: The API requires the client to delete a Pokemon Region using name
      parameters:
        - name:
        - in: query
          schema:
            type: string
      responses:
        200:
          description: 'Pokemon Region was successfully added'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PokemonRegionDTO"
              examples:
                example-1:
                  $ref: "#/components/schemas/PokemonRegionDTO"
        400:
          description: 'Pokemon Region validation empty exception'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PokemonRegionDTO"
              examples:
                example-1:
                  $ref: "#/components/examples/EmptyRegionExample"
                example-2:
                  $ref: "#/components/examples/MinimumPopulationExample"

components:
  schemas:
    PokemonRegionDTO:
      type: object
      properties:
        name:
          type: string
          example: Kanto
        population:
          type: integer
          example: 100
  examples:
    EmptyRegionExample:
      value:
        name: must not be blank
        population: must not be null
    MinimumPopulationExample:
      value:
        name: must not be blank
        population: must be greater than or equal to 10